Several 64bit fixes to the model and things code

+++ doomsday/engine/portable/include/p_object.h
@@ -31,7 +31,7 @@
 #endif
 
 // This macro can be used to calculate a thing-specific 'random' number.
-#define THING_TO_ID(mo) ( (mo)->thinker.id * 48 + ((unsigned)(mo)/1000) )
+#define THING_TO_ID(mo) ( (long)(mo)->thinker.id * 48 + ((unsigned long)(mo)/1000) )
 
 // We'll use the base mobj template directly as our mobj.
 typedef struct mobj_s {
